+++ NEWS.md
@@ -1,3 +1,8 @@
+# AzureAuth 1.3.0
+
+- Allow obtaining tokens for the `organizations` and `consumers` generic tenants, in addition to `common`.
+- More robust handling of expiry time calculation for AAD v2.0 authentication.
+
 # AzureAuth 1.2.5
 
 - Change maintainer email address.